+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2022-10-XX
+
+## Added
+
+- Added the `extend()` function, which serializes and appends an iterable of Python objects to a UTF-8-encoded jsonl file.
+
+## Changed
+
+- Changed the `append()` function to serialize and append a single Python object to a UTF-8-encoded jsonl file.
+
 ## [0.2.2] - 2022-12-24
 
 ### Fixed